lib/intel_allocator: Check validity of the file descriptor
All procedures for opening particular types of allocators
query the gtt size. Even if the user has provided the
boundaries of the address space managed by the allocator,
the query is performed for the sake of validation.

gem_aperture_size() called with an invalid id returns a
default value instead of failing. Thus, we need to
additionally check whether the file descriptor is valid.

Signed-off-by: Andrzej Turko <andrzej.turko@linux.intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>