lib/i915/submission: Rework gem_test_all_engines to use intel_ctx_t (v3)
This function used to be called gem_test_engine but it's only ever used
to test all the engines in the GPU so we rename it for the one thing
it's used for.  We also convert to intel_ctx_t.  It now creates a new
context with all physical engines and iterates over all the engines in
it.

v2 (Ashutosh Dixit):
 - Delete the now non-existent @engine param from the docs

v3 (Jason Ekstrand):
 - Create the context after re-opening the device

Signed-off-by: Jason Ekstrand <jason@jlekstrand.net>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Acked-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>