lib: Ensure the spinner reuses the canonical address
For resubmitting a spinner, we pinned the objects into their addresses
used for the initial submission (so that we can remove the relocations).
When supplying a pinned address, the kernel insists that it should be in
canonical form, so let's also make sure that it is on setting the pin
flag.

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>