tests/kms_cdclk: Fix CDCLK failure condition
Driver still works in a such way that we never reduce CDCLK, but always
only increase it. That was done in order to avoid continuous CDCLK switching,

So, in case if new CDCLK is 307200 and there is no change in CDCLK beacuse
of above stated driver optimization; pass the test else check lt condition.

TODO: Enhance this test to lower modes for all connectors in order to lower
      CDCLK and get a proper DVFS behavior, along with the above mentioned
      kernel change.

Cc: Uma Shankar <uma.shankar@intel.com>
Signed-off-by: Swati Sharma <swati2.sharma@intel.com>
Reviewed-by: Uma Shankar <uma.shankar@intel.com>