i915: Improve the precision of command parser checks
The previous gem_has_cmdparser helper took an engine and did nothing
with it.  We delete the engine parameter and use the general helper for
the ALL_ENGINES cases.  For cases where we really do care about
something more precise, we add a version which takes an intel_ctx_cfg_t
and an engine specifier and is able to say whether or not that
particular engine has the command parser enabled.

Signed-off-by: Jason Ekstrand <jason@jlekstrand.net>
Cc: Lakshminarayana Vudum <lakshminarayana.vudum@intel.com>
Acked-by: Daniel Vetter <daniel.vetter@ffwll.ch>