tests/kms_plane_alpha_blend: Skip coverage-vs-premult-vs-none for 6bpc panels
Intel platforms using 6bpc panels have dithering ON
and show CRC mismatch in coverage-vs-premult-vs-none
Doing a crc based test with dithering is not a great idea as pixels
will change due to the 8bit to 6bits truncation with dithering.

This patch skips the subtest for 6bpc panels. Currently, there is no better
way of fetching the bpc info other than EDID or display_info. Currently using
i915_display_info for the same. Hence restricting the check to i915 devices only.

Credits-to: Uma Shankar <uma.shankar@intel.com>
Credits-to: Juha-pekka Heikkila <juha-pekka.heikkila@intel.com>
Credits-to: Modem Bhanuprakash <Bhanuprakash.Modem@intel.com>
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
Signed-off-by: Vidya Srinivas <vidya.srinivas@intel.com>