test/i915/gem_ringfill: Create contexts on the right FD with NEW_FD
We originally did the re-open dance because we were copying engines from
ctx0 on one FD to ctx0 on the other and we needed both FDs.  When we
converted to intel_ctx_t in 81e16a59d788 ("tests/i915/gem_ringfill:
Convert to intel_ctx_t"), we got the FDs wrong in the context create.
This both fixes the bug and avoids the whold FD dance.

Signed-off-by: Jason Ekstrand <jason@jlekstrand.net>
Fixes: 81e16a59d788 ("tests/i915/gem_ringfill: Convert to intel_ctx_t")
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>