tests/kms_invalid_dotclock: Modify the test for bigjoiner
Currently the test assumes that the modes with clock more than the
maximum dot clock will be rejected. This fails in case of
platforms that combine multiple pipes and modes higher than the
maximum dotclock can still be supported.

This patch modifies the test to bump the clock for the given mode
by twice if bigjoiner is supported.

Cc: Manasi Navare <manasi.d.navare@intel.com>
Signed-off-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>
Signed-off-by: Juha-Pekka Heikkilä <juha-pekka.heikkila@intel.com>
Signed-off-by: Patnana Venkata Sai <venkata.sai.patnana@intel.com>
Reviewed-by: Manasi Navare <manasi.d.navare@intel.com>
Signed-off-by: Patnana Venkata Sai <venkata.sai.patnana@intel.com>