lib/core: Add igt_debug_on* variants
If a test calls a function which can return a failure, only the fact
that the function failed can be reported to the caller, with no details
on what actually failed inside.  It could be helpful if functions which
perform several steps or iterations emitted debug messages with failure
details.

To simplify coding and avoid code duplication in functions, igt_debug
variants similar to igt_warn_on and friends could be useful.  Add them.

Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Acked-by: Petri Latvala <petri.latvala@intel.com>