tests/kms_busy: Limit the execution to single connector
As part of reducing the CI time, we restricted the execution
to two pipes (first & last), which also introduced running the
tests on all connectors.

	commit 030c28d099a5c6793c92e2b22c8e9abd3d340fdc
	tests/kms_busy: Limit the execution to two pipes

With that change in multi-display configuration, we run the test
multiple times on the same pipe with different connectors, which
increases the CI time.

This patch will fix the logic to run the tests on a single
connector only.

V2:
* Fix the crash in no-display configs

Cc: Karthik B S <karthik.b.s@intel.com>
Signed-off-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>