test/kms_frontbuffer_tracking: Fix multidraw test
Multidraw tests was not notifying driver about the area it uses
blitter to pain the primary background color between drawing methods.
Without this notification driver will not update pipe and the CRC will
not update causing the tests fail due a CRC mismatch.

All the other callers of fill_fb_region() are safe, doing a page flip
or a atomic commit afterwards.

Reviewed-by: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>
Cc: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>