tests/kms_atomic: Initialize array values before checking them
Currently, the subtest plane-immutable-zpos fails due to a lack of
initialization of a set of arrays allocated in the stack. Later on in
the test, there is a check if some of the array components are NULL
which is always false because of the random values in the memory. This
commit fixes this issue by initializing all arrays with zero at the
beginning of the plane_immutable_zpos function, and now it passes for
amdgpu.

Cc: Harry Wentland <harry.wentland@amd.com>
Cc: Nicholas Choi <Nicholas.Choi@amd.com>
Cc: Sean Paul <seanpaul@chromium.org>
Cc: Mark Yacoub <markyacoub@chromium.org>
Cc: Melissa Wen <melissa.srw@gmail.com>
Cc: Arkadiusz Hiler <arek@hiler.eu>
Cc: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Rodrigo Siqueira <Rodrigo.Siqueira@amd.com>
Reviewed-By: Nicholas Choi <Nicholas.Choi@amd.com>