tests/kms_ccs: Test bad params for all CCS planes
Test creating an FB with incorrect CCS plane parameters for all CCS
planes. While at it move all the incorrect FB parameter subtests earlier
before generating an FB. For these subtests we won't modeset and so
don't need to generate FBs with a test pattern. Exceptions are the
bad-rotation subtests which also do a modeset.

Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>