tests/kms_prime: Create the exporting BO with smem
On i915, to avail the dmabuf, the shared object needs to migratable
into smem. So if the shared object is lmem, then it needs to have
smem as second placement option.

Currently kms_prime sharing the dumb buffer between the devices.
But dumb buffer can't have the placements and resides at lmem for the
dgfx. Hence to meet the i915 expectation for dgfx, we create the BO
using the gem_create with smem as second placement option.

v2: Used gem_mmap__device_coherent for mmaped ptr (Ashutosh)
v3: dumb buffer ioctls are used for non i915 drivers

Signed-off-by: Ramalingam C <ramalingam.c@intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com> (v2)
Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>