tests/i915/gem_cs_tlb: Drop gem_require_ring() check
80f48722f429 ("tests/i915/gem_cs_tlb: Convert to intel_ctx_t") passes a
intel_ctx_t (initialized with intel_ctx_create_all_physical()) into
run_on_ring(). However gem_require_ring() checks if the ring is available
to context 0, not the intel_ctx_t which is passed in. This will cause the
test to skip for all engines not available to context 0. Hence drop the
gem_require_ring() check since we already know that the intel_ctx_t has all
physical engines available to it.

Cc: "Zbigniew Kempczynski <zbigniew.kempczynski@intel.com>"
Fixes: 80f48722f429 ("tests/i915/gem_cs_tlb: Convert to intel_ctx_t")
Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Reviewed-by: Arjun Melkaveri <arjun.melkaveri@intel.com>