tests/kms_psr2_su: Skip test when necessary
The whole test needs to skip in display 13 platforms as
PSR2_SU_STATUS is not being updated anymore by HW.
It was not reliable in older display versions and now it is 100% dead.

For display 12 and newer platforms now we have PSR2 selective fetch
enabled default, so software is doing all the tracking of the areas
that needs updates but we still can't do that for frontbuffer
rendering so we are doing full frame fetches for now what will cause
this test to fail, so also skipping it for the frontbuffer subtest.

Cc: Jeevan B <jeevan.b@intel.com>
Cc: Swati Sharma <swati2.sharma@intel.com>
Cc: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>
Reviewed-by: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>