tests/kms_writeback.c: fix the `fill_fb` function
Currently, the `fill_fb` function, instead of filling the framebuffer
with the whole pixel, is only filling it with the rightmost byte.

This is happening because, even though the memset accepts an int as
a parameter, it's casting the value to unsigned char and will fill
the memory with this 8 bits value.

If we setup the pixel to 0xff0000|f0|, the buffer will filled with:

| addr     | = 0f
| addr + 1 | = 0f
| addr + 2 | = 0f
| addr + 3 | = 0f
    ...
| addr + N | = 0f

This patch address this issue by manually copying the entire integer.

Signed-off-by: Igor Torrente <igormtorrente@gmail.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>