lib/igt_fb/tgl+: Prevent CCS FB AUX surface size overalignment
So far the page alignment of the size of CCS AUX surfaces was ensured
by aligning the height of the surface to 64, which multiplied by the 64
byte CCS tile width resulted in the multiple of a 4k value. This can
overallocate the CCS surface up to ~1/4th of one main surface tile-row
size. Avoid the overallocation by simply aligning the size to 4k.

Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>