intel-ci: blacklist gem_exec_alignment@pi-shared
The reason for this is that it doesn't work with the dma-resv-object locking introduced
recently in the code,and it's not going to be fixed anytime soon.

drm/i915: Perform execbuffer object locking as a separate step
To help avoid evicting already resident buffers from the batch we're
processing, perform locking as a separate step.

Link:https://cgit.freedesktop.org/drm-tip/commit/?id=5cd57f676bb946a00275408f0dd0d75dbc466d25

Signed-off-by: Mastan Katragadda <mastanx.katragadda@intel.com>
Reviewed-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>