lib/igt_fb: fix the plane_size array for yuv formats
clear_yuv_buffer() checks the size of the plane_size[] to
make sure its greater than the number of planes to avoid
overflows.

The plane_size[] is fixed to two currently.

However some of the formats like YV12 indeed have more than
2 planes in the format_desc[] hence this incorrectly failing
this check.

Increase the size of the plane_size[] to match the correct
max number of planes.

changes in v3:
	- use drm_format->num_planes to create the plane_size
	- add assert checks for num of planes before memset

Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>