lib/igt_core: Ensure subtest result is skip if all dynamic subtests skip
Even though the best practices for dynamic subtests is to not execute
the dynamic subtest at all instead of skipping, there are a couple of
tests that to the opposite, without an easy way to construct them
along best practices. Accomodate this with fixing the handling of
skipping dynamic subtests so that the result of the containing subtest
becomes a skip also if all dynamic subtests skip, not just when there
are no dynamic subtests.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Arkadiusz Hiler <arek@hiler.eu>
Cc: Jari Tahvanainen <jari.tahvanainen@intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>