tests/kms_big_fb: Check plane capabilities properly in async flip tests
Currently IVB only supports 64bpp formats on the sprite planes
but not on the primary plane. This means
igt_display_has_format_mod(64bpp)->true and thus we will try to
run the 64bpp async flip tests. But since that test uses the
primary plane we will fail to actually enable the plane. Fix
this by making sure the plane actually supports the format (and
let's also check the rotation for good measure).

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>