lib: Introduce typed cleanups
We close devices on exit but that doesn't work when inside
igt_subtest_group is used igt_skip(), as it uses longjmp,
causing the code to go out of scope and miss close().

Start introducing standard types with automatic cleanup courtesy of
gcc's __attribute__((cleanup)). This also works with clang, when
using volatile integers.

As an example, we start with an fd that will automatically call
close() on going out of scope, and crucially before atexit where
we will want to check for resource leaks.

[mchehab: add test descriptions]
Suggested-by: Andrzej Hajda <andrzej.hajda@intel.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Andrzej Hajda <andrzej.hajda@intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>
Acked-by: Nirmoy Das <nirmoy.das@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>