tests/kms_async_flips: Fix the crc test vblank request
The test fails to populate the horrible vblank ioctl
flags correctly unless it happens to be running on the
first crtc. Use the helper we have to cook up the
necessary magic flags.

Should make the test work on some CHV machines that have
to use pipe C due to their restricted pipe->port routing
capabilities.

Reviewed-by: Karthik B S <karthik.b.s@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>